+++ src/main/java/com/portfolio/backend/repository/EducationRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface EducationRepository extends JpaRepository <Education, Long> {
+public interface EducationRepository extends JpaRepository <Education, Integer> {
     public Optional<Education> findByDegree (String degree);
     public boolean existsByDegree(String degree);
 }
