+++ src/main/java/com/portfolio/backend/repository/ExperienceRepository.java
@@ -4,7 +4,7 @@ import com.portfolio.backend.model.Experience;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ExperienceRepository extends JpaRepository <Experience, Long> {
+public interface ExperienceRepository extends JpaRepository <Experience, Integer> {
     public Optional<Experience> findByCompany (String company);
     public boolean existsByCompany(String company);
 }
