+++ src/main/java/com/portfolio/backend/repository/ProjectRepository.java
@@ -4,7 +4,7 @@ import com.portfolio.backend.model.Project;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProjectRepository extends JpaRepository <Project, Long>{
+public interface ProjectRepository extends JpaRepository <Project, Integer>{
     public Optional<Project> findByNameProject (String nameProject);
     public boolean existsByNameProject(String nameProject);
 }
