+++ src/main/java/com/portfolio/backend/repository/SkillRepository.java
@@ -4,7 +4,7 @@ import com.portfolio.backend.model.Skill;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SkillRepository extends JpaRepository <Skill, Long>{
+public interface SkillRepository extends JpaRepository <Skill, Integer>{
     public Optional<Skill> findByNameSkill (String nameSkill);
     public boolean existsByNameSkill(String nameSkill);
 }
